Code cleanup and comments
